# Conflicts:
#	src/main/java/ru/kata/spring/boot_security/demo/controller/AdminController.java
#	src/main/java/ru/kata/spring/boot_security/demo/controller/UserController.java
#	src/main/java/ru/kata/spring/boot_security/demo/entity/Role.java
#	src/main/java/ru/kata/spring/boot_security/demo/entity/User.java
#	src/main/java/ru/kata/spring/boot_security/demo/repository/RoleRepository.java
#	src/main/java/ru/kata/spring/boot_security/demo/repository/UserRepository.java
#	src/main/java/ru/kata/spring/boot_security/demo/service/RoleService.java
#	src/main/java/ru/kata/spring/boot_security/demo/service/RoleServiceImpl.java
#	src/main/java/ru/kata/spring/boot_security/demo/service/UserService.java
#	src/main/java/ru/kata/spring/boot_security/demo/service/UserServiceImpl.java
#	src/main/resources/templates/index.html
#	src/main/resources/templates/user.html

+++ src/main/java/ru/kata/spring/boot_security/demo/controller/AdminController.java
@@ -27,7 +27,6 @@ public class AdminController {
     public AdminController(UserServiceImpl userService, RoleServiceImpl roleService) {
         this.userService = userService;
         this.roleService = roleService;
-
     }
 
     @GetMapping
@@ -52,7 +51,7 @@ public class AdminController {
     }
 
     @GetMapping("/edit/{id}")
-    public String edit(Model model, @PathVariable("id") int id) {
+    public String edit(Model model, @PathVariable("id") Long id) {
         model.addAttribute("user", userService.getUserById(id));
         model.addAttribute("roles", roleService.getRoleList());
         return "edit";
@@ -67,7 +66,7 @@ public class AdminController {
     }
 
     @DeleteMapping("/delete/{id}")
-    public String delete(@PathVariable("id") int id) {
+    public String delete(@PathVariable("id") Long id) {
         userService.deleteUser(id);
         return "redirect:/admin";
     }
