+++ src/main/java/ru/kata/spring/boot_security/demo/service/UserService.java
@@ -9,11 +9,11 @@ public interface UserService {
 
     void updateUser(User updateUser);
 
-    void deleteUser(int id);
+    void deleteUser(Long id);
 
     User findByUserName(String username);
 
-    User getUserById(int id);
+    User getUserById(Long id);
 
     List<User> getAllUsers();
